bản mới

+++ core/domain/src/main/java/com/example/domain/model/Weather.kt
@@ -1,6 +1,6 @@
 package com.example.domain.model
 
 data class Weather(
-    val temp: Float,
+    val temp: Int,
     val type: String
 )
